requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.12.0
PyYAML>=6.0.1
Jinja2>=3.1.0
python-dotenv>=1.0.0
markdown>=3.5.0
python-frontmatter>=1.0.0